確認ダイアログの実装

+++ frontend/src/pages/Article/Article.vue
@@ -3,8 +3,21 @@
 <script>
 import axios from 'axios'
 import Article from '@/models/article'
+import Vue from 'vue'
+import VuejsDialog from 'vuejs-dialog'
+// VuejsDialogのデフォルトCSSをimport(モーダル用)
+import 'vuejs-dialog/dist/vuejs-dialog.min.css'
+
+// VuejsDialogを使用することをvueに通知
+Vue.use(VuejsDialog)
 
 const DOMAIN_BASE = process.env.DOMAIN_BASE
+const message = '本当に削除しますか？'
+let options = {
+  type: 'basic',
+  okText: 'はい',
+  cancelText: 'いいえ'
+}
 
 export default {
   props: ['id'],
@@ -21,13 +34,21 @@ export default {
       })
   },
   methods: {
-    deleteArticle: function () {
-      axios.delete(`${DOMAIN_BASE}articles/${this.id}`)
-        .then(response => {
-          console.log(response.data)
-          console.log(response.status)
+    // 削除前に実行確認のダイアログを表示する
+    dialog: function () {
+      this.$dialog.confirm(message, options)
+        .then(() => {
+          axios.delete(`${DOMAIN_BASE}articles/${this.id}`)
+            .then(response => {
+              this.$router.push('/articles')
+              console.log(response.status)
+              console.log('削除が実行されました')
+            })
         })
-    }
+        .catch(function () {
+          console.log('キャンセルされました')
+        })
+    },
   }
 }
 </script>
